Bug 1313327 - Flatten the Telemetry Event enums. r=janerik
To support a non-templated API for recording events by enum, all the enum
values must be of the same type. Inheritence doesn't really seem to be a thing
in C++ enums, so that means flattening them all into a single, big enum the
way Scalars and Histograms work.

Alas.

Differential Revision: https://phabricator.services.mozilla.com/D18691

UltraBlame original commit: bcbe1dc287257a75b5435322f351722875920d94

+++ toolkit/components/telemetry/core/TelemetryEvent.cpp
@@ -99,7 +99,8 @@ namespace TelemetryIPCAccumulator = mozilla::TelemetryIPCAccumulator;
 
 namespace {
 
-const uint32_t kEventCount = mozilla::Telemetry::EventID::EventCount;
+const uint32_t kEventCount =
+      static_cast<uint32_t>(mozilla::Telemetry::EventID::EventCount);
 
 
 const uint32_t kExpiredEventId = std::numeric_limits<uint32_t>::max();
